Moving the editmode test to the HDRLatest project, since the previous was only used for playmode tests
