+++ TestProjects/AlembicHDRPLatest/Assets/Data~/CrashAsset.abc
