+++ TestProjects/AlembicHDRPLatest/Assets/Tests/EditorModeTests.cs
@@ -13,5 +13,27 @@ namespace UnityEditor.Formats.Alembic.Tests
             Assert.IsTrue(true);
             yield return new ExitPlayMode();
         }
+
+        [Test]
+        public void CorruptAlembicFileDoesNotCrashInEditor()
+        {
+            var fileSource = "Assets/Data~/CrashAsset.abc";
+            //generate unique name, as previous import failures might prevent the import process to run.
+            var fileDestination = $"Assets/CrashAsset-{UnityEngine.GUID.Generate().ToString()}.abc";
+
+            try
+            {
+                AssetDatabase.StartAssetEditing();
+                FileUtil.CopyFileOrDirectory(fileSource, fileDestination);
+                AssetDatabase.StopAssetEditing();
+                AssetDatabase.ImportAsset(fileDestination);
+            }
+            finally
+            {
+                AssetDatabase.DeleteAsset(fileDestination);
+            }
+
+            Assert.Pass("Editor has not crashed.");
+        }
     }
 }
