--- TestProjects/AlembicStandaloneBuild/Assets/EditModeTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dc37d7eed390b49b194a0f45f0be1ced
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
