--- TestProjects/AlembicStandaloneBuild/Assets/EditModeTests/EditModeTests.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0db01041096f4e8fb257f13a37500bdc
-timeCreated: 1769452954
