--- TestProjects/AlembicStandaloneBuild/Assets/EditModeTests/Unity.Formats.Alembic.UnitTests.Runtime.Standalone.asmdef
@@ -1,23 +1,0 @@
-{
-    "name": "Unity.Formats.Alembic.UnitTests.Editor.Standalone",
-    "rootNamespace": "",
-    "references": [
-        "UnityEditor.TestRunner",
-        "Unity.Formats.Alembic.Runtime"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
