--- TestProjects/AlembicStandaloneBuild/Assets/EditModeTests/Unity.Formats.Alembic.UnitTests.Runtime.Standalone.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9b2bade7266734e5db463675f0395458
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
